requirements.txt
Django==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==0.19.2
psycopg2==2.9.11
